+++ src/main/java/com/faithjoyfundation/autopilotapi/v1/dto/car_managment/CarRequest.java
@@ -20,7 +20,7 @@ public class CarRequest {
 
     @NotNull
     @Positive
-    private Double mileage;
+    private Integer mileage;
 
     @NotNull
     @Positive
